energy_service/requirements.txt
fastapi
uvicorn
httpx
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
deprecated